Split integration test test_signing_in_and_my_stuff into test-login-failures and test-my-stuff
